LayoutTests/http/tests/misc/DOMContentLoaded-event-expected.txt

ALERT: DOMContentLoaded event fired on document!
ALERT: DOMContentLoaded event bubbled to window!
ALERT: load event fired!
ALERT: PASSED
Test case for bug 5122: Equivalent of Mozilla's DOMContentLoaded needed

You should see a PASSED message.
